Relocate recipe-difficulty.enum.ts to shared folder
