+++ server/src/common/enums/enums.ts
@@ -1,5 +1,5 @@
 export {ApiRoutes} from './api/api'
 export {AuthRoutes} from './auth/auth'
 export {DbTablesNamesEnum} from './db/db-tables-names'
+export {RecipeRoutes, RecipeDifficulty} from "./recipe/recipe"
 export {ENV} from './env/env'
-export {RecipeDifficulty} from "./recipeDifficulty/recipeDifficulty"
