+++ server/src/common/enums/recipe/recipe.ts
@@ -1,1 +1,1 @@
-export {RecipeRoutes} from 'shared/common/enums/enums'
+export {RecipeRoutes, RecipeDifficulty} from 'shared/common/enums/enums'
