--- server/src/common/enums/recipeDifficulty/recipeDifficulty.ts
@@ -1,1 +1,0 @@
-export {RecipeDifficulty} from "./recipeDifficulty.enum"
