+++ shared/common/enums/enums.ts
@@ -1,3 +1,3 @@
 export {AuthRoutes} from './auth/auth'
 export {ApiRoutes} from './api/api'
-export {RecipeRoutes} from './recipe/recipe'
+export {RecipeRoutes, RecipeDifficulty} from './recipe/recipe'
