+++ shared/common/enums/recipe/recipe.ts
@@ -1,1 +1,2 @@
 export {RecipeRoutes} from './recipe-routes.enum'
+export {RecipeDifficulty} from './recipe-difficulty.enum'
